Moved process_ex_ants to metrics_io, broke up tests
