+++ hera_qm/metrics_io.py
@@ -703,3 +703,37 @@ def load_metric_file(filename):
 
     _recursively_validate_dict(metric_dict)
     return metric_dict
+
+
+def process_ex_ants(ex_ants=None, metrics_file=None):
+    """
+    Return list of excluded antennas from command line argument.
+    Input:
+       ex_ants -- comma-separated value list of excluded antennas
+       metrics_file -- file containing array info from hera_qm
+    Output:
+       list of excluded antennas
+    """
+
+    # test that there are ex_ants to process
+    if ex_ants is None and metrics_file is None:
+        return []
+    else:
+        xants = []
+        if ex_ants is not None:
+            if ex_ants != '':
+                for ant in ex_ants.split(','):
+                    try:
+                        if int(ant) not in xants:
+                            xants.append(int(ant))
+                    except ValueError:
+                        raise AssertionError(
+                            "ex_ants must be a comma-separated list of ints")
+        if metrics_file is not None:
+            metrics = load_metric_file(metrics_file)
+            xants_m = metrics["xants"]
+            for ant in xants_m:
+                ant_num, pol = ant
+                if ant_num not in xants:
+                    xants.append(int(ant_num))
+        return xants
