+++ hera_qm/tests/test_metrics_io.py
@@ -413,3 +413,24 @@ def test_read_old_all_string_ant_json_files():
     # This function recursively walks dictionary and compares
     # data types together with nt.assert_type_equal or np.allclose
     qmtest.recursive_compare_dicts(test_metrics_new, test_metrics_old)
+
+def test_process_ex_ants_empty():
+    ex_ants = ''
+    xants = metrics_io.process_ex_ants(ex_ants=ex_ants)
+    nt.assert_equal(xants, [])
+
+def test_process_ex_ants_string():
+    ex_ants = '0,1,2'
+    xants = metrics_io.process_ex_ants(ex_ants=ex_ants)
+    nt.assert_equal(xants, [0, 1, 2])
+
+def test_process_ex_ants_bad_string():
+    ex_ants = '0,obvious_error'
+    nt.assert_raises(AssertionError, metrics_io.process_ex_ants, ex_ants=ex_ants)
+
+def test_process_ex_ants_string_and_file():
+    ex_ants = '0,1'
+    met_file = os.path.join(DATA_PATH, 'example_ant_metrics.hdf5')
+    xants = metrics_io.process_ex_ants(ex_ants=ex_ants, metrics_file=met_file)
+    nt.assert_equal(xants, [0, 1, 81])
+    return
