+++ hera_qm/tests/test_utils.py
@@ -404,21 +404,3 @@ def test_dynamic_slice():
     nt.assert_equal(b.shape, (1, 5))
 
     nt.assert_raises(ValueError, utils.dynamic_slice, 'foo', slice(0, None))
-
-def test_process_ex_ants():
-    ex_ants = ''
-    xants = utils.process_ex_ants(ex_ants=ex_ants)
-    nt.assert_equal(xants, [])
-
-    ex_ants = '0,1,2'
-    xants = utils.process_ex_ants(ex_ants=ex_ants)
-    nt.assert_equal(xants, [0, 1, 2])
-
-    ex_ants = '0,obvious_error'
-    nt.assert_raises(AssertionError, utils.process_ex_ants, ex_ants=ex_ants)
-
-    ex_ants = '0,1'
-    met_file = os.path.join(DATA_PATH, 'example_ant_metrics.hdf5')
-    xants = utils.process_ex_ants(ex_ants=ex_ants, metrics_file=met_file)
-    nt.assert_equal(xants, [0, 1, 81])
-    return
