+++ hera_qm/utils.py
@@ -659,38 +659,3 @@ def dynamic_slice(arr, slice_obj, axis=-1):
     for ax in axis:
         slices[ax] = slice_obj
     return arr[tuple(slices)]
-
-
-def process_ex_ants(ex_ants=None, metrics_file=None):
-    """
-    Return list of excluded antennas from command line argument.
-    Input:
-       ex_ants -- comma-separated value list of excluded antennas
-       metrics_file -- file containing array info from hera_qm
-    Output:
-       list of excluded antennas
-    """
-    from . import metrics_io
-
-    # test that there are ex_ants to process
-    if ex_ants is None and metrics_file is None:
-        return []
-    else:
-        xants = []
-        if ex_ants is not None:
-            if ex_ants != '':
-                for ant in ex_ants.split(','):
-                    try:
-                        if int(ant) not in xants:
-                            xants.append(int(ant))
-                    except ValueError:
-                        raise AssertionError(
-                            "ex_ants must be a comma-separated list of ints")
-        if metrics_file is not None:
-            metrics = metrics_io.load_metric_file(metrics_file)
-            xants_m = metrics["xants"]
-            for ant in xants_m:
-                ant_num, pol = ant
-                if ant_num not in xants:
-                    xants.append(int(ant_num))
-        return xants
